use only 16 bits of timers
even if the timer is 32 bits, use only bottom 16 bits for consistency

+++ Mcu/f421/Inc/blutil.h
@@ -70,7 +70,7 @@ static inline void bl_timer_disable(void)
     BL_TIMER->ctrl1_bit.tmren = FALSE;
 }
 
-static inline uint32_t bl_timer_us(void)
+static inline uint16_t bl_timer_us(void)
 {
     return BL_TIMER->cval;
 }
